LAB 4 | task 1.2
Added constructor checks that input data is a two-dimensional array and
stores it in the internal field @data, which is frozen to ensure
immutability.
Added get_element method allows you to get an element by specified
indices, after checking that the indices are in the allowed range.
Added row_count and column_count methods return the number of rows and
columns respectively